+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.7.32
+
+* `LockedDataset` will now throw `WouldWaitMore`